+++ demo/api/src/db/fixtures/generators/draft-js-migration-page-fixture.service.ts
@@ -79,6 +79,15 @@ const legacyDraftJsContent = {
         },
         {
             key: "55555555-5555-5555-5555-555555555555",
+            text: "Nested bullet point",
+            type: "unordered-list-item",
+            depth: 1,
+            inlineStyleRanges: [],
+            entityRanges: [],
+            data: {},
+        },
+        {
+            key: "55555555-5555-5555-5555-5555555555aa",
             text: "Second bullet point",
             type: "unordered-list-item",
             depth: 0,
@@ -91,6 +100,15 @@ const legacyDraftJsContent = {
             text: "First ordered item",
             type: "ordered-list-item",
             depth: 0,
+            inlineStyleRanges: [],
+            entityRanges: [],
+            data: {},
+        },
+        {
+            key: "66666666-6666-6666-6666-6666666666aa",
+            text: "Nested ordered item",
+            type: "ordered-list-item",
+            depth: 1,
             inlineStyleRanges: [],
             entityRanges: [],
             data: {},
